jobstest: move FakeResumer to jobstest
Release note: None
