+++ pkg/jobs/jobs_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobsprotectedts"
+	"github.com/cockroachdb/cockroach/pkg/jobs/jobstest"
 	"github.com/cockroachdb/cockroach/pkg/keyvisualizer"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts/ptpb"
@@ -207,7 +208,7 @@ func noopPauseRequestFunc(
 	return nil
 }
 
-var _ jobs.TraceableJob = (*jobs.FakeResumer)(nil)
+var _ jobs.TraceableJob = (*jobstest.FakeResumer)(nil)
 
 func (rts *registryTestSuite) setUp(t *testing.T) {
 	rts.ctx = context.Background()
@@ -261,7 +262,7 @@ func (rts *registryTestSuite) setUp(t *testing.T) {
 	rts.onPauseRequest = noopPauseRequestFunc
 
 	jobs.RegisterConstructor(jobspb.TypeImport, func(job *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			TraceRealSpan: rts.traceRealSpan,
 			OnResume: func(ctx context.Context) error {
 				t.Log("Starting resume")
@@ -1151,7 +1152,7 @@ func TestRegistryLifecycle(t *testing.T) {
 		resumerJob := make(chan *jobs.Job, 1)
 		jobs.RegisterConstructor(
 			jobspb.TypeBackup, func(j *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-				return jobs.FakeResumer{
+				return jobstest.FakeResumer{
 					OnResume: func(ctx context.Context) error {
 						resumerJob <- j
 						return nil
@@ -2248,7 +2249,7 @@ func TestShowJobWhenComplete(t *testing.T) {
 	defer close(done)
 	jobs.RegisterConstructor(
 		jobspb.TypeImport, func(_ *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-			return jobs.FakeResumer{
+			return jobstest.FakeResumer{
 				OnResume: func(ctx context.Context) error {
 					select {
 					case <-ctx.Done():
@@ -2412,7 +2413,7 @@ func TestJobInTxn(t *testing.T) {
 		},
 	)
 	jobs.RegisterConstructor(jobspb.TypeBackup, func(job *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				t.Logf("Resuming job: %+v", job.Payload())
 				atomic.AddInt32(&hasRun, 1)
@@ -2452,7 +2453,7 @@ func TestJobInTxn(t *testing.T) {
 		},
 	)
 	jobs.RegisterConstructor(jobspb.TypeRestore, func(job *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(_ context.Context) error {
 				return errors.New("RESTORE failed")
 			},
@@ -2550,7 +2551,7 @@ func TestStartableJobMixedVersion(t *testing.T) {
 	require.NoError(t, err)
 
 	jobs.RegisterConstructor(jobspb.TypeImport, func(job *jobs.Job, settings *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{}
+		return jobstest.FakeResumer{}
 	}, jobs.UsesTenantCostControl)
 	var j *jobs.StartableJob
 	jobID := jr.MakeJobID()
@@ -2591,7 +2592,7 @@ func TestStartableJob(t *testing.T) {
 		return func() { resumeFunc.Store(prev) }
 	}
 	jobs.RegisterConstructor(jobspb.TypeRestore, func(job *jobs.Job, settings *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				return resumeFunc.Load().(func(ctx context.Context) error)(ctx)
 			},
@@ -2777,7 +2778,7 @@ func TestStartableJobTxnRetry(t *testing.T) {
 	defer s.Stopper().Stop(ctx)
 	jr := s.JobRegistry().(*jobs.Registry)
 	jobs.RegisterConstructor(jobspb.TypeRestore, func(job *jobs.Job, settings *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{}
+		return jobstest.FakeResumer{}
 	}, jobs.UsesTenantCostControl)
 	rec := jobs.Record{
 		Details:  jobspb.RestoreDetails{},
@@ -2819,7 +2820,7 @@ func TestRegistryTestingNudgeAdoptionQueue(t *testing.T) {
 	defer jobs.ResetConstructors()()
 	resuming := make(chan struct{})
 	jobs.RegisterConstructor(jobspb.TypeBackup, func(_ *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				resuming <- struct{}{}
 				return nil
@@ -2905,7 +2906,7 @@ func TestMetrics(t *testing.T) {
 	fakeBackupMetrics := makeFakeMetrics()
 	jobs.RegisterConstructor(jobspb.TypeBackup,
 		func(j *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-			return jobs.FakeResumer{
+			return jobstest.FakeResumer{
 				OnResume: func(ctx context.Context) error {
 					defer fakeBackupMetrics.N.Inc(1)
 					return waitForErr(ctx)
@@ -2919,7 +2920,7 @@ func TestMetrics(t *testing.T) {
 	)
 
 	jobs.RegisterConstructor(jobspb.TypeImport, func(_ *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				return waitForErr(ctx)
 			},
@@ -3139,7 +3140,7 @@ func TestLoseLeaseDuringExecution(t *testing.T) {
 	defer jobs.ResetConstructors()()
 	resumed := make(chan error, 1)
 	jobs.RegisterConstructor(jobspb.TypeBackup, func(j *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				defer close(resumed)
 				_, err := s.InternalExecutor().(isql.Executor).Exec(
@@ -3209,7 +3210,7 @@ func TestPauseReason(t *testing.T) {
 	defer close(done)
 	resumeSignaler := newResumeStartedSignaler()
 	jobs.RegisterConstructor(jobspb.TypeImport, func(job *jobs.Job, settings *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				resumeSignaler.SignalResumeStarted()
 				select {
@@ -3464,7 +3465,7 @@ func TestPausepoints(t *testing.T) {
 	defer s.Stopper().Stop(ctx)
 	idb := s.InternalDB().(isql.DB)
 	jobs.RegisterConstructor(jobspb.TypeImport, func(job *jobs.Job, settings *cluster.Settings) jobs.Resumer {
-		return jobs.FakeResumer{
+		return jobstest.FakeResumer{
 			OnResume: func(ctx context.Context) error {
 				if err := registry.CheckPausepoint("test_pause_foo"); err != nil {
 					return err
@@ -3604,7 +3605,7 @@ func TestJobTypeMetrics(t *testing.T) {
 
 	for typ := range typeToRecord {
 		jobs.RegisterConstructor(typ, func(job *jobs.Job, _ *cluster.Settings) jobs.Resumer {
-			return jobs.FakeResumer{
+			return jobstest.FakeResumer{
 				OnResume: func(ctx context.Context) error {
 					<-ctx.Done()
 					return ctx.Err()
